# api/openapi.yaml
openapi: 3.0.3
info:
  title: DevSecOps Jira Dashboard API
  description: API for the DevSecOps Jira Dashboard - Overview, Projects, and Settings modules
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/projects:
    get:
      tags:
        - Projects
      summary: Get projects list
      description: Returns the paginated list of projects with filters. Each project includes its repositories when expanded.
      parameters:
        - name: period
          in: query
          required: false
          description: Time period filter for the projects list
          schema:
            type: string
            enum:
              - last_week
              - last_month
              - last_3_months
              - last_6_months
              - last_year
            default: last_month
        - name: search
          in: query
          required: false
          description: Free text search on project name
          schema:
            type: string
            example: "Xenon"
        - name: status
          in: query
          required: false
          description: Comma-separated list of status IDs to filter by. Empty means all.
          schema:
            type: string
            example: "active,at_risk"
        - name: client
          in: query
          required: false
          description: Comma-separated list of client IDs to filter by. Empty means all.
          schema:
            type: string
            example: "abn_amro,barclays"
        - name: specialization
          in: query
          required: false
          description: Comma-separated list of specialization IDs to filter by. Empty means all.
          schema:
            type: string
            example: "devsecops,devops"
        - name: offset
          in: query
          required: false
          description: Number of items to skip for pagination
          schema:
            type: integer
            default: 0
            example: 0
        - name: limit
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by
          schema:
            type: string
            enum:
              - project
              - onboarded_date
            default: project
        - name: sort_order
          in: query
          required: false
          description: Sort direction
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsListSuccessResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidParameterError'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonServerError'
  /api/v1/projects/action:
    post:
      tags:
        - Projects
      summary: Perform action on a project
      description: Performs an action on a project based on the action field. Supports "mark_not_applicable" (raises Jira ticket) and "mark_complete" (updates status to Completed).
      operationId: performProjectAction
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectActionRequest'
      responses:
        '200':
          description: Action performed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectActionSuccessResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidParameterError'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonServerError'
...
  /api/v1/sync/servicenow/projects:
    post:
      tags:
        - Sync
      summary: Sync projects from ServiceNow
      description: Receives project data from ServiceNow and creates or updates project records in the database. Each request contains a single project with its associated repositories and metadata.
      operationId: syncServiceNowProjects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncProjectRequest'
      responses:
        '200':
          description: Project sync completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncSuccessResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidParameterError'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '409':
          description: Sync already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncConflictError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      required:
        - status_code
        - status
        - message
        - data
      properties:
        status_code:
          type: integer
          description: HTTP status code
        status:
          type: string
          enum: [success, failed, error]
          description: Response status indicator
        message:
          type: string
          description: Response message
        data:
          description: Response data payload

    # Common Server Error Response
    CommonServerError:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            status_code:
              type: integer
              example: 500
            status:
              type: string
              example: "error"
            message:
              type: string
              example: "An unexpected error occurred. Please try again later"
            data:
              type: array
              items:
                type: object
              example: []

    # Unauthorized Error Response
    UnauthorizedError:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            status_code:
              type: integer
              example: 401
            status:
              type: string
              example: "failed"
            message:
              type: string
              example: "Authentication token is missing or expired"
            data:
              type: array
              items:
                type: object
              example: []

    # Invalid Parameter Error Response
    InvalidParameterError:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            status_code:
              type: integer
              example: 400
            status:
              type: string
              example: "failed"
            message:
              type: string
              example: "Invalid query parameter: 'period' must be one of [last_week, last_month, last_3_months, last_6_months, last_year]"
            data:
              type: array
              items:
                type: object
              example: []

    # ===========================
    # Overview Response Schemas
    # ===========================
...
    ProjectsListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            status_code:
              type: integer
              example: 200
            status:
              type: string
              example: "success"
            message:
              type: string
              example: "Projects retrieved successfully"
            data:
              type: object
              properties:
                projects:
                  type: array
                  items:
                    $ref: '#/components/schemas/ProjectItem'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    ProjectItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the project
          example: "550e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          description: Name of the project
          example: "Project Xenon"
        client_name:
          type: string
          nullable: true
          description: Name of the client associated with the project
          example: "ABN AMRO"
        project_type:
          type: string
          description: Type of the project
          example: "App"
        onboarded_date:
          type: string
          format: date
          description: Date when the project was onboarded
          example: "2026-02-20"
        repository_count:
          type: integer
          description: Number of repositories under this project
          example: 3
        status:
          type: string
          description: Current status of the project
          example: "COMPLETED"
        overdue_days:
          type: integer
          description: Number of days the project is overdue (0 if not at risk)
          example: 5
        repositories:
          type: array
          description: List of repositories under this project
          items:
            $ref: '#/components/schemas/RepositoryItem'
        not_applicable_details:
          nullable: true
          description: Details from the Jira ticket when the project status is "Not Applicable". Null when the project has any other status.
          allOf:
            - $ref: '#/components/schemas/NotApplicableDetails'
    NotApplicableDetails:
      type: object
      description: Jira ticket details associated with a project marked as Not Applicable. Retrieved from the jira_tickets table by project ID.
      properties:
        reason_category:
          type: string
          nullable: true
          description: Category of the reason for marking the project as not applicable
          example: "Client Restriction"
        comments:
          type: string
          nullable: true
          description: Comments explaining why the project was marked as not applicable
          example: "Project Wavefront has been marked as not applicable due to client restrictions."
        evidence_url:
          type: string
          nullable: true
          description: URL to the evidence file uploaded when marking the project as not applicable
          example: "https://s3.amazonaws.com/bucket/evidence/file.pdf"
        type:
          type: string
          nullable: true
          description: Type of the Jira ticket
          example: "Not Applicable"
        priority:
          type: string
          nullable: true
          description: Priority level of the Jira ticket
          example: "Medium"
        assignee:
          type: string
          nullable: true
          description: Person assigned to the Jira ticket
          example: "john.doe@example.com"
        jira_status:
          type: string
          nullable: true
          description: Current status of the Jira ticket
          example: "Done"
    RepositoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the repository
          example: "880e8400-e29b-41d4-a716-446655440030"
        name:
          type: string
          description: Name of the repository
          example: "xenon-infra-core"
        onboarded_date:
          type: string
          format: date
          description: Date when the repository was onboarded
          example: "2026-02-27"
        status:
          type: string
          description: Current status of the repository
          example: "PASSED"
    Pagination:
      type: object
      properties:
        offset:
          type: integer
          description: Number of items skipped
          example: 0
        limit:
          type: integer
          description: Items per page
          example: 10
        total_items:
          type: integer
          description: Total number of items
          example: 44

    # Project Action Request
    ProjectActionRequest:
      type: object
      required:
        - project_id
        - action
      properties:
        project_id:
          type: string
          format: uuid
          description: UUID of the project to perform the action on
          example: "550e8400-e29b-41d4-a716-446655440001"
        action:
          type: string
          enum:
            - mark_not_applicable
            - mark_complete
          description: Action to perform on the project
          example: "mark_not_applicable"
        reason_category:
          type: string
          description: Reason category (required for mark_not_applicable)
          example: "Client Restriction"
        comments:
          type: string
          description: Comments explaining the reason (required for mark_not_applicable)
          example: "Project Wavefront has been marked as not applicable due to client restrictions."
        evidence_file:
          type: string
          format: binary
          description: Evidence file attachment (optional, for mark_not_applicable)

    # Project Action Success Response
    ProjectActionSuccessResponse:
      type: object
      required:
        - status_code
        - status
        - message
      properties:
        status_code:
          type: integer
          example: 200
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Action performed successfully"

    # Not Found Error Response
    NotFoundError:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            status_code:
              type: integer
              example: 404
            status:
              type: string
              example: "failed"
            message:
              type: string
              example: "Project not found"
            data:
              type: array
              items:
                type: object
              example: []

    # ===========================
    # Repository Detail Schemas
    # ===========================
...
    SyncProjectRequest:
      type: object
      required:
        - sn_project_id
        - project_name
        - onboarded_date
        - project_type
      properties:
        sn_project_id:
          type: string
          description: ServiceNow project identifier (stored in projects.sn_project_id)
          example: "SN-PRJ-00456"
        project_name:
          type: string
          description: Name of the project
          example: "Payments Platform"
        onboarded_date:
          type: string
          format: date
          description: Date when the project was onboarded
          example: "2026-03-15"
        project_type:
          type: string
          description: Type/category of the project
          example: "Infrastructure"
        is_applicable:
          type: boolean
          description: Whether the project is applicable for DevSecOps onboarding
          default: true
          example: true
        client:
          type: string
          description: Client name associated with the project
          example: "ABN AMRO"
        approver:
          type: string
          description: Name or email of the approver for the project
          example: "barath.uthamasathish@zeb.co"

    # --- DevSecOps Tickets Sync from ServiceNow ---
...
    SyncSuccessResponse:
      type: object
      required:
        - status_code
        - status
        - message
      properties:
        status_code:
          type: integer
          example: 200
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Sync completed successfully"
    SyncConflictError:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            status_code:
              type: integer
              example: 409
            status:
              type: string
              example: "failed"
            message:
              type: string
              example: "A sync operation is already in progress. Please try again later."
            data:
              type: array
              items:
                type: object
              example: []

    # ===========================
    # Settings Schemas
    # ===========================
...
security:
  - BearerAuth: []